refactor: api 패키지, view 패키지 분리
