+++ src/main/java/cart/controller/GlobalExceptionHandler.java
@@ -3,6 +3,7 @@ package cart.controller;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import cart.service.LoginFailException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpHeaders;
