+++ src/main/java/cart/controller/api/ProductCreateController.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.api;
 
 import javax.validation.Valid;
 import java.net.URI;
