+++ src/main/java/cart/controller/api/ProductSearchController.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.api;
 
 import java.util.List;
 import java.util.stream.Collectors;
