+++ src/main/java/cart/controller/view/CartViewController.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
