+++ src/main/java/cart/controller/view/ProductViewController.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.view;
 
 import java.util.List;
 import java.util.stream.Collectors;
