+++ src/main/java/cart/controller/view/UserViewController.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.view;
 
 import java.util.List;
 import java.util.stream.Collectors;
