+++ src/test/java/cart/controller/api/CartSearchControllerTest.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.api;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -9,6 +9,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import java.util.List;
 
+import cart.controller.AbstractProductControllerTest;
 import cart.domain.cart.Cart;
 import cart.domain.product.Product;
 import cart.domain.user.User;
