+++ src/test/java/cart/controller/api/ProductDeleteControllerTest.java
@@ -1,8 +1,9 @@
-package cart.controller;
+package cart.controller.api;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import cart.controller.AbstractProductControllerTest;
 import org.junit.jupiter.api.Test;
 
 @SuppressWarnings("NonAsciiCharacters")
