+++ src/test/java/cart/controller/api/ProductRequestControllerTest.java
@@ -1,8 +1,9 @@
-package cart.controller;
+package cart.controller.api;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import cart.controller.AbstractProductControllerTest;
 import cart.dto.ProductRequest;
 import org.junit.jupiter.api.Test;
